+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.11.2
 
 - Treat "00" as the high (or single) value in table ranges as "100" (assuming it is a percentage roll).
+- PF1: When the statblock-library module is enabled, the name of the creature in the statblock in the Actor's Notes tab will be a link to the matching compendium entry.
 
 ## 2.11.1